rename LandingPage to Landing to solve stupid OSX casing bug
